ADD , created HomePage Folder and moved related files in it for better organisation
