+++ my-app/src/App.js
@@ -1,12 +1,11 @@
-import Cart from './Cart';
-import FoodList from './FoodList';
-import ItemComponent from './ItemsComponent';
+import FoodList from './HomePage//FoodList';
+import ItemComponent from './HomePage/ItemsComponent';
 
 import {BrowserRouter,Route,Routes,Link,} from 'react-router-dom'
 import Navbar from './Navbar';
 
 import Home from './Home';
-import CartItem from './cartItem';
+import CartItem from './HomePage/cartItem';
 import Categories from './Categories';
 
 
@@ -15,15 +14,6 @@ function App() {
   return (
     <div className="App">
       <Navbar/>
-      
-      {/* <Link to='/Categories'>Categories</Link> */}
-     {/* {<BrowserRouter>
-      <main>
-        <Routes>
-          <Route path='/' element={<Home/>}/>
-          <Route path= '/Categories' element={<Categories/>}/>
-        </Routes>
-      </main> */}
       
         <Routes>
           <Route path='/' element = {<Home/>} />
