+++ my-app/src/HomePage/ItemsComponent.js
